Name the binary file
The crate is currently defined as a library crate and Cargo.toml
does not define any binaries. I think this makes it so that the binary
is compiled to target/release/main. This fix explicitly defines a binary
with the name genome_parser, that goes to target/release/genome_parser.